(GH-2)(GH-15) Make Blockquotes Show Appropriately
Adds styling to make blockquotes show appropriately. A snippet of js
has also been added that will style any blockquote with the warning
emoji with a yellow border instead of the standard grey. This snippet
can be added to to detect other emojis, and color those blockquotes
different colors as well.