(GH-2) Code Block Highlighting/Prism/Clipboard
Uses a newly created Prism.js file that includes only the languages
needed in our application. This also includes a plugin for adding line
numbers and a copy to clipboard function.

Since the native plugin for prism.js includes a copy to clipboard
feature now, the custom clipboard action that was added with
Clipboard.js has been removed.

A new custom scss file has been added as well to support the syntax
highlighting theme.

This follows what is seen on chocolatey.org.