(GH-2)(GH-10) Minimize/Expand Left Side Navigation
These changes allow all nested pages to be shown in the left side
navigation. The pages are minimized (collapsed) unless the page is
navigated to. In this case, the element they are located inside will
be expanded.

This contains no custom styling, and focuses on functionality. Styling
will be added later.