(GH-2)(GH-6) Remove Search Box Placeholder
A search box was added into the UI as part of the initial style
overhaul, however at this point it is not functional. For now, it is
being removed.

To add back in later, the source code should be searched for "TODO:
Search", and carry out whatever those notes say to do.