(GH-44) Add an Alert Notification Bar
Adds an alert notification bar at the top of the page. This can be used
for important messages we want the user to see. When the "x" is
clicked, an item is set in the users session storage which will remove
the alert for their current session.

If the text inside the alert is removed, the alert will not be shown.