(GH-45) Fix right sidebar links
When a heading contains a link, it renders incorrectly in the sidebar
since we are essentially trying to add a link into a link, which doesn't
work. This inline module strips the raw text from a heading and returns
that, which then does result in valid links in the right sidebar. At some
point, we could split this into it's own standalone module, but an issue
has been raised to have this go directly into Statiq.Framework, at which
point this workaround could be removed:

https://github.com/statiqdev/Statiq.Framework/issues/142