(GH-52) Make Cookie Work Across All Domains
The cookie set by the cookie banner notice has been set to have the
domain of chocolatey.org. This will allow the cookie to be read and
used on the main chocolatey.org and any subdomain.

If the cookie is set on a local environment, it will not be set with
this property. This has been done so that the cookie still works and
can be tested while running locally. The cross domain
functionality will only work on the actual live site.