(GH-6) Enhance Search on Mobile
Previously on mobile, the height of the search modal was set with css
`min-height: 100vh`. This is causing a few problems as this property
does not account for the addition of the address bar and other items
added to the browser when the user scrolls up or down. In return, this
is hiding some of the bottom of the modal.

The alternative that has been added is configuring the height with some
JS, which is much more accurate and accommodates for any added entities
by the browser.

Solution:
https://css-tricks.com/the-trick-to-viewport-units-on-mobile/

Also, as a test/fix `debug` has been set to `true` on the search
configuration. This was done because when a user presses "done" on
mobile (after using the search box), focus is taken away, and the
search results disappear. This will potentially allow the results to
stay and allow the user click (tap) them.