(GH-58) Set an Expiration Date on Cookie
This adds an expiration date to the cookie notice so that the cookie
persist through multiple user sessions, actually 100 years. This is
inline with how it is set on chocolatey.org.

Without setting an expiration date, the cookie is reset every time the
user starts a new session.

A slight update to the cookie notice button has been made to stay
inline with chocolatey.org and make the button slightly larger on
mobile, making it easier to tap.