(GH-6) Adjust Search for Mobile
Right now, on mobile, when trying to search, the search box will
disappear right as the user tries to scroll the page. This is due to a
resize event that is being triggered by an added scroll bar.

A fix has been put in that will prevent this. A similar problem is
happening with the "Jump to Section" button, which this will fix as
well.

A few small styling updates have been added for mobile as well.