(theme) Add Responsiveness to Video iframes
This wraps each iframe in an embed div that will help to scale the
video on any viewport. A custom viewport has been added since the
videos in the GUI docs are at a non-standard size.

For future embed implementations, different classes may need to be
used based on the video ratio aspects.

For more information see:
https://getbootstrap.com/docs/4.5/utilities/embed/