(GH-83) Update front matter
The RedirectFrom section of the FrontMatter is used from redirecting
from one page to another. Statiq will create an empty redirect page at
the location in the attribute, and point it at the current page. In this
case, we don't want/need a redirect, as we are splitting, and leaving an
existing page.

The Order element in the Front Matter controls the page order within
that folder. Starting at 70 doesn't make sense, so I have changed this
to 10. The only time this is different is for the index.md file, where
the order applies parent elements in the tree, so leaving that as 60
will ensure that the api node remaining in the same place.