(layout) Add Header and Main Elements to Layout
The top navigation and global navigation/alert banner have been added
into a header element. The rest of the body has been moved into a main
element. While it looks like a lot of changes, this is due to items now
being nested inside these elements. This was done to stay consistent
with the blog and allow easier usage of choco-theme classes/styles/js.