(GH-88) Add Styling to Custom 404 Page
There have been changes made to the main layout to support pages that
do not contain a sidebar. If the page does not contain a sidebar, and
should be a full page width, then the front matter `NoSidebar: true`
should be added to said page.

This adds custom styles to the 404 page to make it fit better with the
theme, and keep consistent with the blog.