(GH-90) Updated numbering of rules
Due to the fact that rules can change type, using a numbering structure
to identify the rule type is the wrong approach. If a requirement
changes to a guideline, or vice versa, that number would give a invalid
indication of what the rule is.

Rather, increment all the rules after each other.