(maint) Full sync with generated docs
The changes in this commit are as a direct result of running the
GernerateDocs.ps1 script in the chocolatey/choco repository, and
copying the output into the correct location within this repository.
The changes are mostly cosmetic, i.e. white space changes, and
formatting, but going forward, it should only be actual changes from
the source code that are copied into these files.