(ui) Provide Tab Shell for UI
While providing a tabbed interface completely in markdown isn't
possibly with the current set up, the tab pane content area is. This
provides a shell tab area for content to be added in as necessary.

Currently, there is a known issue where any headers inside the tabbed
area are added into the right hand navigation, but are not navigatable
unless the selected tab pane is already open.