(#264) Correct casing for Win Defender example
This commit updates the example for Windows Defender when configuring
a generic virus scanner, to use the same case as in other parts of the documentation.
Additionally, the quoting of the value have been updated to ensure that
passed to the generic virus scanner will work when the path contains spaces.

This change is necessary to be consistent with the documentation,
as well as only using upper case letters currently do not work for the
replacement variable.