(doc) Replace Gitter refs with Community Chat
As we have replaced Gitter with Discord the links have to be updated. Rather than specifying Discord, I've used a generic Community Chat term. This means, if we ever change away from Discord in future, we don't have to update the docs again. The link being used is a short url we can also update outside of the docs. Gitter rooms will be hidden so any references to chats have also been removed.