(#349) Update Top Navigation Structure
While the top navigation looks the same, the structure of it has
changed slightly. When adding the correct JS files and CSS classes, a
sticky navigation can be enabled.

This sticky navigation feature is mostly for the main chocolatey.org,
but this required changes to be made to all sites since they
share common CSS via choco-theme.