Add information around deprecations
It is possible, when a config/feature is deprecated and removed between
package versions of Chocolatey, that the config/featute values remain
in the chocolatey.config file.

This is by design, as it means that the configuration file is backwards
compatible, especially when going between Chocolatey Licensed and back
to OSS.

This is not immediately obvious though, so we have to call it out.