(#364) Adjust classes from choco-theme update
Work on choco-theme has been done to generate utility classes in a more
efficient manner. In this process, a few class names were modified so
that they correlated better with the style they were providing.
This repository has been skimmed for these classes and updated
accordingly. In addition, a few inline styles have been removed and
replaced with a class.