(#367) Navigate to nested items on hash change
The implementation for navigating to a hash that located inside of a
collapsed or tabbed element already existed, however it was commented
out in choco-theme. This has been updated and is now enabled for use.

On the install page, a user will now correctly be navigated when
clicking on a link in the right hand nav under "More Install Options".