(#384) Upgrade to Yarn 3.1.1
Yarn v3.1.1 is now being utilized. Yarn v3.1.1 works a bit differently
in that it is not installed globally, but per project basis. This means
there can still be a global v1 version on a machine, but this
individual project can use v3.1.1. This is due to how Yarn now takes
advantage of `yarnPath` via the `.yarnrc.yml` file.

More information on `yarnPath`:
https://yarnpkg.com/configuration/yarnrc#yarnPath

More information on how Yarn is managed going forward:
https://yarnpkg.com/getting-started/install