(#367) Navigate to nested items on hash change
This updates how anchors are navigated to when they are located
inside a collapsed element. The previous implementation stopped working,
and this is a long term solution to fix the bug for good. This is
mostly used on the docs site on the setup Chocolatey CLI page, and on
the pricing page, however it can be used anywhere needed.

Changes done via choco-theme:
https://github.com/chocolatey/choco-theme/commit/b9572ff012c7bc3dd6024dddb70bc92f30a86f31