(doc) Add info about Maintenance and Support
Added a new fragment with information about Mainteance and Support for
licensed customers, and how this can impact them when we release a new
version of Chocolatey that is no longer compatible with older version
of the Chocolatey Components.

This has direct impacts on some customers that may be using a perpetual
license.

This fragment has been added to the release notes pages for Chocolatey,
Chocolatey Licensed Extension, and the Chocolatey GUI Licensed
Extension, but it can be added to others as required.