(#491) Set Cookie on Mobile for Cookie Banner
Recently a handful of JavaScript functions were rewritten to remove
jQuery. In the process, the script that controlled setting of the
cookie when a user clicks the "I accept" button in the cookie banner
broke. It has been adjusted slightly to take into account multiple
buttons, which allows the cookie to be set on both desktop and mobile.