(#669) Ensure Callouts Render Properly
This updates choco-theme to bring in changes to ensure that all
callotus are rendered properly. Previously, if a callout contained
anything other than a `<p>` element on the first line, this would throw
a JS error. Now, everything renders as it should and looks great.

In addition, a few callouts have been updated to bring them up to date
with the current choco-theme formatting.