(#329) Improve readability of CCM install docs
An update to choco-theme allowed for the creation of a tab structure in
the docs and we have utilized them to streamline and improve
the CCM installation documentation. Each package installation document
will be updated with the tab structure to facilitate easier reading of
the Database Authentication Scenarios, with each scenario being given
its own tab. Some of the links I originally created were direct links instead of
xrefs, so I fixed this after it was brought to my attention.