(#706) Add specific share images for release notes
This adds in new parameters that can be added to the frontmatter of
any markdown page, "TwitterImage" and "OgImage". These should be links
to an image that is meant to be used when the link is shared on social
media. The "TwitterImage" parameter is used when the link is shared on
Twitter, and the "OgImage" parameter is used when the link is shared on
Facebook or most other platforms like Discord or Slack.

The "TwitterImage" utilizes the "summary_large_image" type, which
requires an image ratio of 2:1. The "OgImage" must be 1200x630 pixels.