(#716)(#731) Enhance content and styles of home
This adds a new concept of "Highlights" to the home page. A Highlight is
something that is new or important to the user that has recently been
changed in the docs. The README.md file has been updated with
information on how to add a new Highlight.

In addition the the concept of Highlights, the look and feel as well as
the content has been massively overhauled. The home page now features
cards with links to the top 5-6 visited pages in our docs according to
Google Analytics. Below the cards are a few additional useful links as
well as information about how to find support and join the
Chocolatey Community.

Most of the content that was previously on the home page has been
removed, as it could be found in many other places in the docs as it
stands, and wasn't being utilized by users. Now instead of overwhelming
users that come to our docs, we give them the most relevant information
to what they are looking for.

It is important to keep updating the Highlights section with future
noteworthy changes, and also updating the top 5-6 most visited links.

In addition, Bootstrap has been pinned to version 5.2.3, to avoid
automatically upgrading to Bootstrap 5.3.0 which has a breaking change.