(doc) Add information about constraint warning
This change provides information about a constraint warning
that may be outputted to the user when they upgrade their packages
or install new packages that have the potential of breaking version ranges.

Such a warning is usually shown when the package that is upgraded is
a dependency for multiple other packages. But may happen it other
circumstances as well.