(#817) Add missing redirects
To ensure that we catch all the possible URL's to these now deleted
pages, we need to ensure that we redirect from both the
`quickdeployment`, and `quickdeployment\` variants of the URL. For this
to work with Statiq, we need to be specific about the latter, and
include the full URL to the page, i.e. the thing that the browser/client
do for us as default page. That means that we need something like
`quickdeployment/index.html`.

(#817)Change firewall rules redirect

Signed-off-by: Ryan Richter <ryan@chocolatey.io>