(#856) Update branding documentation
Prior to this change, the docs for applying branding to the
Chocolatey GUI application called out the need for the
Chocolatey GUI licensed extension in such a way as that it
could be missed.

This change adds a note on how to internalize this package
so you can host it on your own repository, and also adds
the package as a dependency inside of the nuspec file used
to create the branding packae. This shifts the burden of
ensuring the licensed extension is installed to choco
which will never forget...because it can't.