(#885) Tutorial - Prepare Packaging Workstation
Prior to this we didn't have a prescriptive set of documentation for preparing
a workstation for the purpose of developing your own Chocolatey packages.

With this commit we now have such documentation that sets up your workstation with
everything needed to work through the tutorials on Package Creation available in
our documentation.