(#959) Use Playwright generated images
This updates all the images used in the Chocolatey Central Management
documentation to use ones that have been auto generated by Playwright.

All automated images are in the `ccm-playwright` folder, while any manual
images have been moved to a new `ccm-manual` folder. This is so
everything is better organized. The `ccm-playwright` folder is intended
to be completely replaced whenever a new run of Playwright has been
done.