(#979) Remove Gulp
This removes Gulp and introduces a number of new
technologies. The build process is now handled by
esbuild, SASS, and a number of new technologies.
Instead of running `gulp` to compile assets, the
command `yarn run choco-theme` is now used. By
running this, all necessary files are put in place
and CSS is purged.