Ensure correct URLs are in xref/sidebar navigation
With the switch to Astro, all urls now end in a
trailing slash. This is a known change, however,
the sidebar navigation was not updated to reflect
that. Now, the sidebar navigation will correctly
link to the correct URL which will avoid the
browser redirecting to the correct URL. This change
also updates Xref links automatically as well. This
will help Google Search Console results and SEO.

An update to most the redirects has also been made
to ensure that they will go to the correct url as
well.