(#1013) Add Hash to CSS and JS files
This adds the commit hash to the CSS and JS files
to ensure that the browser always fetches the
latest version of the files when it is updated.

This sometimes becomes a problem when the CSS
and JS files are cached, doing this will ensure
that these files are updated along with the
content.