(#1021) Rework choco sync command documentation
Previously the documentation for the sync command briefly explained
what it was and how to use it, but left a lot of confusion with customers
attempting to leverage it.

This commit brings additional information about the feature, how to query
for useful information to aid in its use across versions, and provides
clear examples and additonal PowerShell code a customer could use for
automation in their environment.