(doc) Remove incorrect line about extensions
The .extension file extension is a requirement when creating extension
packages. It is this file extension that causes Chocolatey CLI to
extract the files into the extensions folder, in the same way that the
.hook file extension that causes files to be extracted into the hooks
folder, and then the .template file extension that causes files to be
extracted into the templates folder.