(#1178) Use choco-astro
The changes here utilize the new choco-astro package. A "baseline" test
is added here that was previously housed in choco-theme. This tests
performs basic actions to ensure Playwright is set up and running
correclty. The tsconfig.json has been updated slightly to remove options
that are not needed, as they are provided in the base tsconfig that it
is extending from already.