(doc) Add tab completion page
This page is based on the work that was originally created in the
chocolatey/choco repository:

https://github.com/chocolatey/choco/pull/2978

Rather than bring this completion file into the core CLI repository,
the decision was made to add this to the documentation instead. The
thought process here is that additional completions for other terminals
is not something that we will "support", but rather this will be a
community effort.  To that end, creating it as a documentation page
that can be quickly updated is easier than shipping with a new CLI
release.

Co-authored-by: hltdev8642 <hltdev8642@gmail.com>