(maint) Re-ordered release notes
To group all the 1.x and 2.x, and 5.x and 6.x releases together, rather
then intermingling them based solely on the date that they were
released.  The need for this came up during a conversation with Rob
where it was becoming harder and harder to find a specific version in
the release notes.  This was happening both when scrolling the headers
on the right hand side of the page, as well as when looking through the
main content of the document.  By ordering each release based on the
version number makes it much easier to find a specific release.