(#970) Refactored Org Guide to nest pages and include button navigation
I've refactored the Organizational Deployment Guide into a folder/file structure to nest the various components beneath the introduction page. I've included button navigation to go back and forth between pages as well as a button to start from the beginning, and have included a button to navigate to Client Setup once that has been included. This commit also includes minor typo fixes and expansion of acronyms and initialisms.