Added table structure to Nexus doc and refactored variable descriptions
The table structure that was present for the variables on the Jenkins document has been ported over onto the Nexus document for consistency. Additional variables for the Nexus document were added to mimic the variables available on the script. I also rewrote the descriptions for what the variables do in order to maintain consistent voice between the two pages.