(#1233)(#1234)(#1237)(#1241) Bump astro, docs, playwright to 2.1.0
This bumps the versions of @chocolatey-software/astro and
@chocolatey-software/docs and @chocolatey-software/playwright to 2.1.0.
With these version, it includes fixes to first escape characters before
inputting them into DynamicCodeBlocks, adds an option to validate
DynamicCodeBlockInput's, and also adds the ability to hide or disable
tabs in the Tab component. This also updates the Chocolatey Headquarters
address.