(#1232) Rework host-packages and deprecate Chocolatey.Server
This commit reworks the "Host Packages Internally" page within the Features section of the docs in order to provide clear and concise direction for users and outline what does and does not fall under Chocolatey Support.

This commit also deprecates the old Chocolatey.Server option for housing packages that Chocolatey no longer maintains or supports.