(doc) Add information about environment variable
When we shipped CLI 2.5.0, we added a new environment variable called
ChocolateyPreviousPackageVersion, which stores the version number for
the previously installed package version, and this is set when doing a
"choco upgrade".  However, this environment variable never made its way
into the docs site.

This commit addresses that, by adding it into the correct location.